+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Minor improvement about the filtering by polygon
 * Fix an issue to return Py-QGIS-Server version
 
 ## 1.1.1 - 2022-07-29